PythonTraining/基础-实操-项目源码/项目02 Python实战操作源码/Code（实例源码及使用说明）/10/1~14/flask/requirements.txt
alembic==1.0.1
asn1crypto==0.24.0
blinker==1.4
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
Flask==1.0.2
Flask-HTTPAuth==3.3.0
Flask-Mail==0.9.1
Flask-Migrate==2.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
Pillow==5.3.0
pycparser==2.19
PyMySQL==0.9.2
python-dateutil==2.7.3
python-editor==1.0.3
requests==2.22.0
six==1.11.0
SQLAlchemy==1.2.12
urllib3==1.25.2
Werkzeug==0.14.1
WTForms==2.2.1
